backend/requirements.txt
# Core framework
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
python-multipart>=0.0.9

# Database
motor==3.3.1
pymongo==4.5.0

# Data & Finance
yfinance>=0.2.36
pandas>=1.5.0
numpy>=1.23.0
feedparser>=6.0.0

# LLM SDKs — direct, no wrappers
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.5.0

# Utilities
python-dotenv==1.2.1
pydantic>=2.0.0
httpx>=0.24.0
requests>=2.28.0
aiohttp>=3.8.0
anyio>=3.6.0

# Misc
python-dateutil>=2.8.0
pytz>=2023.3
six>=1.16.0
certifi>=2023.0
charset-normalizer>=3.0.0
idna>=3.4
urllib3>=1.26.0
typing_extensions>=4.5.0
annotated-types>=0.4.0
sniffio>=1.3.0

# Authentication
firebase-admin>=6.0.0

# Rate Limiting
slowapi>=0.1.9

# Encryption
cryptography>=42.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Phase 3 dependencies
scipy>=1.10.0
pyotp>=2.9.0

